requirements.txt
Flask
python-dotenv
requests
spotipy
flask_login
SQLAlchemy